requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
pymysql==1.1.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
rapidfuzz==3.5.2
